my attempt
my very basic attempt using javascript. i spent about 2 hours to get this far. Next steps would be to create functions for rotating and moving the ships, then logic for what happens when a ship is lost and how to set a beacon so other ships wont get lost at the same coordinate.